move env_no_snap_for_external_calls to certbot.util
